Added CLI system with commands: `version`, `clear`, `hello`, `routes`, and `pm` for managing packages, routes, and more. Integrated `CLIHandler` and updated `RuntimeHandler` to support command-driven functionality.

+++ awt_src/classes/data/DataManager.php
@@ -242,6 +242,9 @@ class DataManager
 
 
 
+        if(empty($ownerName))
+            return false;
+
         $result = $this->deleteOwnerDirectories($ownerName);
 
         if (!$result)
